abstract medical_info_extractor
